requirements.txt
fastapi
uvicorn
python-dotenv
face_recognition
Pillow
cryptography
numpy
python-multipart
sqlitecloud